requirements.txt
lighter-sdk
python-telegram-bot
python-dotenv
websockets
asyncio
aiohttp
aptos-sdk==0.7.1
